requirements.txt
fastapi
uvicorn[standard]
openai
openai-agents
python-dotenv
pydantic-settings
azure-ai-documentintelligence